# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000/v1"
info:
  title: WASAText
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
...
paths:
  /session:
    post:
      tags: ["Session"]
      operationId: doLogin
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Oggetto contenente il nome utente per effettuare il login
              required:
                - name
              properties:
                name:
                  type: string
                  description: Nome utente per il login
                  minLength: 3
                  maxLength: 16
                  pattern: "^.*?$"
                  example: "Maria"
      responses:
        "201":
          description: User log-in action successful
          content:
            application/json:
              example:
                identifier: "sess_987654321"
              schema:
                type: object
                description: Risposta con l'identificatore di sessione
                required:
                  - identifier
                properties:
                  identifier:
                    type: string
                    description: Token/ID di sessione generato
                    minLength: 1
                    maxLength: 256
                    pattern: "^[a-zA-Z0-9\\-_.]+$"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /groups:
    post:
      tags: ["Group"]
      operationId: createGroup
      summary: Create a WasaText group 
      description: |-
        Crea un nuovo gruppo di WasaText.
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload per creare un gruppo
              required:
                - name
              properties:
                name:
                  type: string
                  description: Name of the group to be created
                  minLength: 1
                  maxLength: 50
                  pattern: "^[a-zA-Z0-9_ ]+$"
                memberIds:
                  type: array
                  description: ID di altri utenti da aggiungere subito al gruppo (opzionale)
                  minItems: 0
                  maxItems: 50
                  items:
                    type: string
                    minLength: 1
                    maxLength: 64
                    pattern: "^[a-zA-Z0-9_\\-]+$"
      responses:
        "201":
          description: Successful creation of WasaText group
          content:
            application/json:
              example:
                id: "grp_001"
                name: "Amici"
                photoUrl: "https://example.com/group.jpg"
              schema:
                $ref: "#/components/schemas/Group"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
    get:
      tags: ["Group"]
      operationId: getGroupDetails
      summary: Retrieve WasaText group details 
      description: |-
        Ottiene i dettagli di un gruppo specifico.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Successful retrieval of group details
          content:
            application/json:
              example:
                id: "grp_001"
                name: "Amici"
                photoUrl: "https://example.com/group.jpg"
              schema:
                $ref: "#/components/schemas/Group"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/members:
    parameters:
      - $ref: "#/components/parameters/groupId"

    get:
      tags: ["Group"]
      operationId: getGroupMembers
      summary: Retrieve WasaText group members
      description: |-
        Ottiene l'elenco dei membri di un gruppo specifico.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Successful retrieval of group members
          content:
            application/json:
              example:
                members:
                  - id: "usr_123"
                    username: "mario_rossi"
                    photoUrl: "https://example.com/photo.jpg"
              schema:
                $ref: "#/components/schemas/GroupMemberListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    post:
      tags: ["Group"]
      operationId: addToGroup
      summary: Add WasaText people in the group 
      description: |-
        Aggiunge membri in un gruppo specifico.
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberIdList"
      responses:
        "200":
          description: Successful retrieval of group members
          content:
            application/json:
              example:
                members:
                  - id: "usr_123"
                    username: "mario_rossi"
                    photoUrl: "https://example.com/photo.jpg"
              schema:
                $ref: "#/components/schemas/GroupMemberListResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/members/me:
    parameters:
      - $ref: "#/components/parameters/groupId"
    delete:
      tags: ["Group"]
      operationId: leaveGroup
      summary: Leave WasaText group 
      description: |-
        Permette all'utente di uscire da un gruppo specifico.
      security:
        - BearerAuth: []

      responses:
        "204":
          description: Successful leave of the group
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/name:
    parameters:
      - $ref: "#/components/parameters/groupId"
    put:
      tags: ["Group"]
      operationId: setGroupName
      summary: Change WasaText group name
      description: |-
        Aggiorna il nome del gruppo.
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupName"

      responses:
        "200":
          description: Successful update of group name
          content:
            application/json:
              example:
                name: "Nuovo Nome Gruppo"
              schema:
                $ref: "#/components/schemas/GroupName"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /groups/{groupId}/photo:
    parameters:
      - $ref: "#/components/parameters/groupId"

    put:
      tags: ["Group"]
      operationId: setGroupPhoto
      summary: Change WasaText group photo
      description: |-
        Aggiorna la foto del gruppo.
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupPhoto"
      
      responses:
        "200":
          description: Successful update of group photo
          content:
            application/json:
              example:
                photoUrl: "https://example.com/new_group_photo.jpg"
              schema:
                $ref: "#/components/schemas/GroupPhoto"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /media:
    post:
      tags: ["Messages"]
      operationId: uploadMedia
      summary: Upload an image
      description: |-
        Carica un'immagine (PNG, JPEG, GIF o WebP, max 8 MB) inviandone i byte
        grezzi nel corpo della richiesta. L'URL restituito può essere usato come
        photoUrl in un messaggio, nella foto profilo o nella foto di un gruppo.
        Il formato viene riconosciuto dal contenuto del file, non dall'header.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        description: Byte grezzi del file immagine
        content:
          image/png:
            schema: { $ref: "#/components/schemas/BinaryImage" }
          image/jpeg:
            schema: { $ref: "#/components/schemas/BinaryImage" }
          image/gif:
            schema: { $ref: "#/components/schemas/BinaryImage" }
          image/webp:
            schema: { $ref: "#/components/schemas/BinaryImage" }
      responses:
        "201":
          description: Immagine caricata correttamente
          content:
            application/json:
              example:
                id: "med_2f1c8a90-1e4b-4a8f-9d31-6b0d0c9f2a11.png"
                url: "/media/med_2f1c8a90-1e4b-4a8f-9d31-6b0d0c9f2a11.png"
              schema:
                $ref: "#/components/schemas/MediaUploaded"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/mediaId"
    get:
      tags: ["Messages"]
      operationId: getMedia
      summary: Retrieve an uploaded image
      description: |-
        Restituisce un'immagine caricata in precedenza. Non richiede
        autenticazione perché viene referenziata da tag img del browser, che non
        inviano l'header Authorization; gli identificativi sono UUID casuali.
      responses:
        "200":
          description: Contenuto binario dell'immagine
          content:
            image/png:
              schema: { $ref: "#/components/schemas/BinaryImage" }
            image/jpeg:
              schema: { $ref: "#/components/schemas/BinaryImage" }
            image/gif:
              schema: { $ref: "#/components/schemas/BinaryImage" }
            image/webp:
              schema: { $ref: "#/components/schemas/BinaryImage" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  schemas:
    Group:
      title: Group
      description: Dettagli completi del gruppo
      type: object
      required:
        - id
        - name
        - photoUrl
      properties:
        id:
          type: string
          description: ID univoco del gruppo
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        name:
          type: string
          description: Nome del gruppo
          minLength: 1
          maxLength: 50
          pattern: "^[a-zA-Z0-9_ ]+$"
        photoUrl:
          type: string
          description: URL della foto di copertina del gruppo (stringa vuota se assente)
          minLength: 0
          maxLength: 2048
          pattern: "^(https?://.+)?$"
    GroupMemberListResponse:
      title: Group Member List
      description: Lista aggiornata dei membri del gruppo.
      type: object
      required: ['members']
      properties:
        members:
          type: array
          description: Elenco degli utenti membri del gruppo
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/User"
    User:
      title: User
      description: Informazioni di base di un utente
      type: object
      required: ['id', 'username']
      properties:
        id:
          type: string
          description: Identificatore univoco utente
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        username:
          type: string
          description: Nome utente univoco
          minLength: 3
          maxLength: 16
          pattern: "^.*?$"
        photoUrl:
          type: string
          description: URL della foto profilo dell'utente (stringa vuota se assente)
          minLength: 0
          maxLength: 2048
          pattern: "^(https?://.+)?$"
...
    BinaryImage:
      title: Binary Image
      description: Byte grezzi di un file immagine (massimo 8 MB)
      type: string
      format: binary
      minLength: 1
      maxLength: 8388608
    MediaUploaded:
      title: Media Uploaded
      description: Riferimento a un'immagine appena caricata
      type: object
      required:
        - id
        - url
      properties:
        id:
          type: string
          description: Identificativo del file, comprensivo di estensione
          minLength: 1
          maxLength: 128
          pattern: "^med_[a-zA-Z0-9\\-]+\\.(png|jpg|jpeg|gif|webp)$"
        url:
          type: string
          description: URL con cui referenziare l'immagine
          minLength: 1
          maxLength: 2048
          pattern: "^/media/[a-zA-Z0-9_.\\-]+$"
...
    MemberIdList:
      title: Member ID List
      description: Elenco degli ID degli utenti da aggiungere al gruppo
      type: object
      required: 
        - memberIds
      properties:
        memberIds:
          type: array
          description: Vettore di stringhe contenente gli ID degli utenti
          minItems: 1
          maxItems: 50
          items:
            type: string
            description: Identificatore univoco di un utente
            minLength: 1
            maxLength: 64
            pattern: "^[a-zA-Z0-9_\\-]+$"
          example: ["user_12345", "user_67890"]
    GroupName:
      title: Group Name
      description: Nome aggiornato del gruppo
      type: object
      required: 
        - name
      properties:
        name: 
          type: string
          description: Nuovo nome del gruppo
          minLength: 1
          maxLength: 50
          pattern: "^[a-zA-Z0-9_ ]+$"
          example: "Friends Group"
    GroupPhoto:
      title: Group Photo
      description: URL aggiornato della foto di profilo del gruppo
      type: object
      required: 
        - photoUrl
      properties:
        photoUrl:
          type: string
          format: url
          description: URL della foto del gruppo
          minLength: 10
          maxLength: 2048
          pattern: "^https?://.+"
          example: "https://example.com/group_photo.jpg"
...
    ErrorResponse:
      title: Error Response
      description: Error response returned by the server in case of failure
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Descrizione dell'errore
          example: "Risorsa non trovata"
...
  responses:
    Unauthorized:
      description: The access token is missing or it's expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource (conversation, message, group, reaction, user) does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |-
        Identificatore utente semplificato (non un JWT reale) ottenuto dall'operazione 'doLogin'.
        Deve essere incluso nell'header 'Authorization' per tutte le richieste protette, nel formato 'Bearer <identifier>'.
  parameters:
...
    groupId: 
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_\\-]+$"
      description: The ID of the group to change photo 
      name: groupId
      in: path
      required: true
    mediaId:
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: "^med_[a-zA-Z0-9\\-]+\\.(png|jpg|jpeg|gif|webp)$"
      name: mediaId
      in: path
      required: true
      description: The ID of the uploaded image
...